% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot_from_data.R
\name{biplot_from_data}
\alias{biplot_from_data}
\title{Produce GGE Biplot From Raw Data}
\usage{
biplot_from_data(genotypes, environments, outcome, all.genotypes = F,
  all.environments = T)
}
\arguments{
\item{genotypes}{vector containing genotype (or option) ID}

\item{environments}{vector containing environment (or context) ID}

\item{outcome}{vector containing numeric outcome values}

\item{all.genotypes}{when not all combinations of GxE exist force biplot to contain all genotypes (defaults to FALSE)}

\item{all.environments}{when not all combinations of GxE exist force biplot to contain all environments (defaults to TRUE)}
}
\description{
Convenience function to transform raw data into format required for GGEBiplotGUI package
}
\examples{
library(agricolae)
data(plrv)
biplot_from_data(plrv$Genotype,plrv$Locality,plrv$Yield)
}
\keyword{GGEBiplot}
\keyword{biplot}
\keyword{wrapper}
