% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{I like the collapse function in paste so much that I am going to make it it's own function
This is probably useful to nobody apart from me}
\usage{
collapse(x, sep = ".")
}
\arguments{
\item{x}{a vector to be collapse}

\item{sep}{collapsing character}
}
\description{
I like the collapse function in paste so much that I am going to make it it's own function
This is probably useful to nobody apart from me
}
