% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.multi.barplot.R
\name{cross.multi.barplot}
\alias{cross.multi.barplot}
\title{Produces a barchart from a multiple response variable with a cross tabulation variable.}
\usage{
cross.multi.barplot(df, crossvar, freq = TRUE, tfreq = "col",
  color = TRUE, crossname = NULL, order = "frequency", tidynames = TRUE,
  multiname = NULL, total = TRUE, ...)
}
\arguments{
\item{df}{data frame with the binary variables}

\item{crossvar}{factor to cross the multiple choices question with}

\item{freq}{use percentages}

\item{tfreq}{type of percentages to compute ("row" or "col")}

\item{color}{Use color scale for cross factor or rely on position only}

\item{crossname}{label for cross factor}

\item{order}{order of responses. "frequency" for frequency order; "alpha" for
alphabetical; "original" for original column order}

\item{tidynames}{tidy the response names to remove . or _ added to variable
names brought in through import or use of split.multi()}

\item{multiname}{label for multiple selection variable}

\item{total}{include total row for row-wise plot}
}
\description{
For column percentages or frequencies the multiresponses are facetted; for row percentages the multiresponses are stacked
Operates in similar way to functions included in questionr, and requires same inputs
}
\examples{
sampledata<-data.frame(Response1=sample(c(0,1),100,replace=TRUE),
Response2=sample(c(0,1),100,replace=TRUE),
Response3=sample(c(0,1),100,replace=TRUE),Category=sample(c("A","B","C"),100,replace=TRUE))
cross.multi.barplot(sampledata[,c("Response1","Response2","Response3")],sampledata$Category,order="original")
cross.multi.barplot(sampledata[,c("Response1","Response2","Response3")],sampledata$Category,order="original",tfreq="row")
cross.multi.barplot(sampledata[,c("Response1","Response2","Response3")],sampledata$Category,order="original",freq=FALSE)
}
\keyword{graph}
\keyword{multiresponse}
