% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heritability.R
\name{heritability}
\alias{heritability}
\title{Calculates heritability statistic(s) from breeding trial. syntax largely stolen from Ric Coe.}
\usage{
heritability(genotypes, environments = NULL, outcome, rep)
}
\arguments{
\item{genotypes}{vector containing genotype (or option) ID}

\item{environments}{vector containing environment (or context) ID}

\item{outcome}{vector containing numeric outcome values}

\item{rep}{rep variable - must be a unique identifier}
}
\description{
Calculates heritability statistic(s) from breeding trial. syntax largely stolen from Ric Coe.
}
\examples{
library(agricolae)
data(plrv)
heritability(plrv$Genotype,plrv$Locality,plrv$Yield,paste(plrv$Locality,plrv$Rep))
}
\keyword{heritability}
