% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luni.R
\name{luni}
\alias{luni}
\title{Number of Unique Values}
\usage{
luni(x, na.rm = T, printvals = F)
}
\arguments{
\item{x}{any R object that can be coerced into a vector}
}
\description{
This function calculates the sample size with option to include or exclude missing values for compatibility with other simple R functions
}
\examples{
y<-c("A","A","B","C","D",NA,"A")
luni(y)
luni(y,na.rm=T)
luni(y,na.rm=F,printvals=T)
}
\keyword{number}
\keyword{of}
\keyword{unique}
\keyword{values}
