% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_errorbar.R
\name{mean_errorbar}
\alias{mean_errorbar}
\title{Plot means with confidence intervals from raw data}
\usage{
mean_errorbar(formula, data, conf.level = 0.95, length = 0.1,
  type = "confidence", las = 2, mfactor = 1, ...)
}
\arguments{
\item{formula}{a formula of the outcome variable and the explanatory variable(s)}

\item{data}{data frame containing variables}

\item{conf.level}{defaults to 0.95. Confidence level for error bars}

\item{length}{width of error bars (in inches) default is 0.1}

\item{type}{type or error bar. Can be "confidence", "prediction", "se" or "sd"}

\item{mfactor}{mulitplication factor for error bars (when type is "se" or "sd")}
}
\description{
Wrapper so that means with error bars can be plotted in a one-liner with same formula input style as boxplot(), lm(), plot() etc.
}
\examples{
library(agricolae)
data(plrv)
mean_errorbar(Yield~Genotype,data=plvr)
}
\keyword{errorbar}
\keyword{mean}
