% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.barplot.R
\name{multi.barplot}
\alias{multi.barplot}
\title{Produces a barchart from a multiple response variable.}
\usage{
multi.barplot(df, order = "frequency", type = "percent", tidynames = TRUE,
  multiname = NULL, fill = "grey", ...)
}
\arguments{
\item{df}{data frame with the binary variables}

\item{order}{order of responses. "frequency" for frequency order; "alpha" for
alphabetical; "original" for original column order}

\item{type}{"percent" or "count"}

\item{tidynames}{tidy the response names to remove . or _ added to variable
names brought in through import or use of split.multi()}

\item{multiname}{label for multiple selection variable}

\item{fill}{Color(s) to use for bars. Must be single color of vector of length of responses}
}
\description{
Produces a barchart from a multiple response variable.
}
\examples{
sampledata<-data.frame(Response1=sample(c(0,1),100,replace=TRUE),
Response2=sample(c(0,1),100,replace=TRUE),
Response3=sample(c(0,1),100,replace=TRUE),Category=sample(c("A","B","C"),100,replace=TRUE))
multi.barplot(sampledata[,c("Response1","Response2","Response3")],order="original")
}
\keyword{graph}
\keyword{multiresponse}
