% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.coincidence.R
\name{multi.coincidence}
\alias{multi.coincidence}
\title{Produces a coincidence matrix plot from a multiresponse variable}
\usage{
multi.coincidence(df, scaleby = "none", diagonal = "sole",
  order = "frequency", text = TRUE, textsize = 3, low = "white",
  high = "red", show = "all", tidynames = TRUE)
}
\arguments{
\item{df}{data frame with the binary variables}

\item{scaleby}{"none" to show counts, "total" to show overall percentages, "row" to show within row percentages, "col" to show within column percentages}

\item{diagonal}{values to place on diagonals. "none" for no values, "sole" for the responses to the category where it was the only response, "all" for all responses to the category}

\item{order}{order of responses. "frequency" for frequency order; "alpha" for
alphabetical; "original" for original column order}

\item{text}{include text labels within cells (TRUE/FALSE)}

\item{textsize}{size for text labels (defaults to 3)}

\item{low}{string indicating of colour to use for lowest value (defaults to "white")}

\item{show}{"all" to show entire matrix, "upper" to show upper half only, "lower" to show lower half only}

\item{tidynames}{tidy the response names to remove . or _ added to variable
names brought in through import or use of split.multi()}

\item{low}{string indicating of colour to use for highest value (defaults to "red")}
}
\description{
Percentages will rarely add up to 100, unless exactly two variables are selected in the multi-response
}
\examples{
sampledata<-data.frame(Response1=sample(c(0,1),100,replace=TRUE),
Response2=sample(c(0,1),100,replace=TRUE),
Response3=sample(c(0,1),100,replace=TRUE),Category=sample(c("A","B","C"),100,replace=TRUE))
multi.coincidence(sampledata[,c("Response1","Response2","Response3")])
multi.coincidence(sampledata[,c("Response1","Response2","Response3")],scaleby="total",diagonal="sole",show="lower")
}
\keyword{graph}
\keyword{multiresponse}
