% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N.R
\name{N}
\alias{N}
\title{Sample Size}
\usage{
N(x, na.rm = F)
}
\arguments{
\item{x}{any R object that can be coerced into a vector}
}
\description{
This function calculates the sample size with option to include or exclude missing values for compatibility with other simple R functions
}
\examples{
y<-c(1,2,3,4,NA)
N(y)
N(y,na.rm=T)
}
\keyword{N}
\keyword{length}
