% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OxCplot.R
\name{OxCPlot}
\alias{OxCPlot}
\title{Standard OxC Interaction plot using ggplot2 so looks nicer than the base R function with minimal effort}
\usage{
OxCPlot(options, contexts, outcome, sort = T, aggregate = "mean",
  optionlab = "", contextlab = "", outcomelab = "", title = "")
}
\arguments{
\item{options}{option/genotype variable making up different lines}

\item{contexts}{context/environment variable making up x axis}

\item{outcome}{vector containing numeric outcome values for y axis}

\item{sort}{sort the x axis from low to high or maintain pre-existing order}

\item{aggregate}{functionn to aggregate data (defaults to mean)}
}
\description{
Plot comparing two treatments tested in multiple identical locations from long format data
}
\examples{
library(agricolae)
data(plrv)
OxCPlot(options=plrv$Genotype,contexts=plrv$Locality,outcome=plrv$Yield)
}
\keyword{Interaction}
\keyword{ggplot2}
\keyword{plot}
