% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskdiff.R
\name{riskdiff}
\alias{riskdiff}
\title{Cumulative Risk plots}
\usage{
riskdiff(outcome, plotID, trtID, grpID = NULL, trt1, trt2, main = "",
  approx = F)
}
\arguments{
\item{outcome}{vector containing numeric outcome values}

\item{plotID}{ID variable for plot/rep/individual}

\item{trtID}{ID variable for treatment}

\item{grpID}{optional ID variable for grouping factor to set colors by}

\item{trt1}{string indicating which treatment to plot on x axis}

\item{trt2}{string indicating which treatment to plot on y axis}

\item{approx}{Add in approximation line (defaults to TRUE)}

\item{same}{TRUE=overlay mutliple curves FALSE= split into 1 plot per group}

\item{splitlab}{string for labelling split variable}
}
\description{
Plot comparing two treatments tested in multiple identical locations from long format data
}
\examples{
riskdiff(outcome=plrv$Yield,
plotID=paste(plrv$Locality,plrv$Rep,sep=""),
trtID=plrv$Genotype,
trt1="Canchan",trt2="Unica",
main="Comparison of Yield: Canchan vs Unica")

riskdiff(outcome=plrv$Yield,
plotID=paste(plrv$Locality,plrv$Rep,sep=""),
trtID=plrv$Genotype,
grpID=plrv$Locality,
trt1="Canchan",trt2="Unica",
main="Comparison of Yield: Canchan vs Unica")
}
\keyword{cumulative}
\keyword{risk}
