% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskplot.R
\name{riskplot}
\alias{riskplot}
\title{Cumulative Risk plots}
\usage{
riskplot(y, splitvar = NULL, approx = T, v = 0, same = T,
  splitlab = "", main = "", xlab = "y")
}
\arguments{
\item{y}{vector containing numeric outcome values}

\item{splitvar}{ID variable for group to split by}

\item{approx}{Add in approximation line (defaults to TRUE)}

\item{same}{TRUE=overlay mutliple curves FALSE= split into 1 plot per group}

\item{splitlab}{string for labelling split variable}
}
\description{
Cumulative Risk plots
}
\examples{
See riskdiff()
}
\keyword{cumulative}
\keyword{risk}
