% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard Error}
\usage{
se(x, na.rm = F)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{remove missing values? (defaults to FALSE)}
}
\description{
This function calculates the naive standard error of a vector
}
\examples{
y<-c(1,2,3,4,NA)
se(y)
se(y,na.rm=T)
}
\keyword{error}
\keyword{se}
\keyword{standard}
