% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Does modified joint regression analysis for variety-by-environment data (P.W. Lane & K. Ryder).}
\usage{
sensitivity(genotypes, environments, outcome, tolerance = 1e-04,
  cex.label = 0.8, ...)
}
\arguments{
\item{genotypes}{vector containing genotype (or option) ID}

\item{environments}{vector containing environment (or context) ID}

\item{outcome}{vector containing numeric outcome values}

\item{tolerance}{tolerance level for convergence}

\item{cex.label}{size of labels on plot}

\item{...}{other arguments to plot()}
}
\description{
Equivalent to RJOINT function in Genstat. Function taken from code included in agridat() package help menu which was not turned into a function (for some reason)
}
\examples{
library(agricolae)
data(plrv)
sensitivity(plrv$Genotype,plrv$Locality,plrv$Yield)
}
\keyword{Sensitivity}
