% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superiority.R
\name{superiority}
\alias{superiority}
\title{Cultivar Superiority}
\usage{
superiority(genotypes, environments, outcome, cex.label = 0.8, ...)
}
\arguments{
\item{genotypes}{vector containing genotype (or option) ID}

\item{environments}{vector containing environment (or context) ID}

\item{outcome}{vector containing numeric outcome values}

\item{cex.label}{size of labels on plot}

\item{...}{other arguments to plot()}

\item{tolerance}{tolerance level for convergence}
}
\description{
Equivalent to GESTABILITY function in Genstat
Calculates stability coefficients for genotype-by-environment data (R.W. Payne).
}
\examples{
library(agricolae)
data(plrv)
superiority(plrv$Genotype,plrv$Locality,plrv$Yield)
}
\keyword{Cultivar}
\keyword{Superiority}
